requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.7.4
grpc-stubs==1.53.0.5
grpcio==1.64.1
grpcio-reflection==1.62.1
grpcio-tools==1.64.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
protobuf==5.27.2
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
python-protobuf==0.0.2
sniffio==1.3.1
types-protobuf==4.25.0.20240410
typing_extensions==4.12.2
